rag-engine/requirements.txt
fastapi
uvicorn[standard]
groq
chromadb
sentence-transformers
python-multipart
python-dotenv
pymupdf